refactor(sidecar): extract shared helpers + WS client + test fakes
Every reference sidecar adapter under
`sdk/python/librefang/sidecar/adapters/` had grown to carry a
near-identical local copy of `_split_message`, `_split_csv`,
`_parse_retry_after`, and (for the 5 WS-using adapters) the entire
hand-rolled RFC 6455 `_WebSocketClient` class. The test suites
were the same story for `_FakeUrlopen`, `_FakeResp`, `_HdrShim`.

A hash audit pre-extraction:

- `_split_message` — 14 files, all behaviour-identical
- `_split_csv` — 7 files, all behaviour-identical
- `_parse_retry_after` — 7 files, identical except for the lower
  clamp (discord 0.1 s, every other adapter 1.0 s)
- `_WebSocketClient` — 5 files: slack/webex/mattermost/qq were
  byte-identical; discord had the same code with extra docstrings.
  Total: ~1900 LOC of duplicated production helpers
- `_FakeUrlopen` / `_FakeResp` / `_HdrShim` — 12 script-driven
  test variants with cosmetic drift (whether the call record
  stored `timeout`, the field name for the decoded body —
  `body_raw` vs `body` — and docstring text). 2 specialised
  variants (gotify status-only, mastodon rotating-reply-id) are
  legitimately different and kept inline.

This commit:

1. Adds `librefang.sidecar.common` exposing `split_message`,
   `split_csv`, `parse_retry_after`. The `parse_retry_after` API
   is `(headers, *, default_secs, floor_secs=1.0, max_secs=60.0)`
   — discord passes `floor_secs=0.1`, every other adapter takes
   the default.
2. Adds `librefang.sidecar.ws` exposing `WebSocketClient` plus
   the RFC 6455 opcode + handshake constants (`OP_TEXT`,
   `OP_PING`, ..., `MAX_FRAME_PAYLOAD`, `WS_GUID`). Lifts
   discord's better-commented version as the canonical body.
3. Adds `tests/_sidecar_fakes.py` with `FakeUrlopen` /
   `FakeResp` / `HdrShim` plus underscore-prefixed back-compat
   aliases so existing call sites keep working without changes.
   The shared `FakeUrlopen` call record carries the **union** of
   fields any pre-extraction variant ever populated (`body_raw`,
   `body`, `params`, `timeout`) so every adapter's assertions
   still find the field it reads.
4. Migrates 16 adapter modules and 12 test files. Each adapter
   keeps a thin `_parse_retry_after(...)` wrapper aliasing the
   shared one (pinned to the adapter's own `MAX_BACKOFF_SECS`)
   so tests that reach for `qq_mod._parse_retry_after` etc. don't
   break.

Verification: `pytest sdk/python/tests/` — 1027 passed in 4.88s.
Zero behaviour change; this is a pure code-location refactor.

Diff summary: 31 files changed, +769 −2479 (net −1710 LOC).

Coverage of the remaining un-migrated channels (matrix, email,
webhook, feishu, dingtalk, wechat, wecom, teams, google_chat,
whatsapp) is left for the in-flight sidecar-migration PR sequence
— each future sidecar imports from the same modules instead of
re-inlining ~250 lines of WS + helpers.
