+++ sdk/python/librefang/sidecar/adapters/bluesky.py
@@ -72,6 +72,7 @@ from typing import Any
 
 from librefang.sidecar import Content, Field, Schema, SidecarAdapter, protocol, run_stdio_main
 from librefang.sidecar import logging as log
+from librefang.sidecar.common import split_message as _split_message
 
 DEFAULT_SERVICE_URL = "https://bsky.social"
 # Bluesky post length cap (graphemes per spec; we approximate with
@@ -91,26 +92,6 @@ SESSION_LIFE_SECS = 5400
 SESSION_REFRESH_BUFFER_SECS = 300
 # Thread-context cache size. 200 entries × ~200 B = ~40 KB, negligible.
 THREAD_CACHE_MAX = 200
-
-
-def _split_message(text: str, max_len: int) -> list[str]:
-    """Chunk `text` into <= max_len pieces, preferring newline splits.
-    Same shape as the ntfy / mastodon / Rust ``split_message`` helper."""
-    if len(text) <= max_len:
-        return [text]
-    chunks: list[str] = []
-    rest = text
-    while len(rest) > max_len:
-        window = rest[:max_len]
-        cut = window.rfind("\n")
-        if cut <= 0:
-            cut = max_len
-        chunks.append(rest[:cut])
-        rest = rest[cut:].lstrip("\n") if cut < max_len else rest[cut:]
-    if rest:
-        chunks.append(rest)
-    return chunks
-
 
 class _LruCache:
     """Tiny fixed-size LRU. OrderedDict-backed: re-insert on get to mark
