+++ sdk/python/librefang/sidecar/adapters/reddit.py
@@ -119,6 +119,7 @@ from typing import Any
 
 from librefang.sidecar import Content, Field, Schema, SidecarAdapter, protocol, run_stdio_main
 from librefang.sidecar import logging as log
+from librefang.sidecar.common import split_message as _split_message
 
 DEFAULT_TOKEN_URL = "https://www.reddit.com/api/v1/access_token"
 DEFAULT_API_BASE = "https://oauth.reddit.com"
@@ -180,26 +181,6 @@ def _normalise_subreddit(value: str) -> str:
     if s.startswith("r/"):
         s = s[2:]
     return s.strip("/")
-
-
-def _split_message(text: str, max_len: int) -> list[str]:
-    """Chunk `text` into <= max_len pieces, preferring newline splits.
-    Matches the Rust ``split_message`` helper used across channels."""
-    if len(text) <= max_len:
-        return [text]
-    chunks: list[str] = []
-    rest = text
-    while len(rest) > max_len:
-        window = rest[:max_len]
-        cut = window.rfind("\n")
-        if cut <= 0:
-            cut = max_len
-        chunks.append(rest[:cut])
-        rest = rest[cut:].lstrip("\n") if cut < max_len else rest[cut:]
-    if rest:
-        chunks.append(rest)
-    return chunks
-
 
 def _parse_reddit_comment(comment: dict, own_username: str) -> dict | None:
     """Parse a Reddit comment JSON object into a ``message`` event.
